+++ smarts/core/sumo_road_network.py
@@ -45,7 +45,7 @@ from .utils.core_math import inplace_unwrap, radians_to_vec, vec_2d
 from .utils.geometry import buffered_shape
 from .utils.glb import make_map_glb, make_road_line_glb
 
-from smarts.core.utils.sumo import sumolib  # isort:skip
+from smarts.core.utils.sumo_utils import sumolib  # isort:skip
 
 
 def pairwise(iterable):